requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
pymongo==4.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0

# LangChain and HuggingFace
langchain==0.0.354
langchain-community==0.0.19
langchain-huggingface==0.0.3
huggingface-hub>=0.23.0
transformers==4.38.2
torch>=2.2.0
sentence-transformers==2.3.1
openai>=1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.0.1
docx2txt==0.8
pandas>=2.2.0
openpyxl==3.1.2
python-pptx==0.6.21
tabulate==0.9.0

# Vector store
chromadb==0.4.23
pydantic>=2.6.0

# Text processing
tiktoken==0.5.2
nltk==3.8.1

# Utilities
numpy>=1.26.0
tqdm==4.66.2
pytest==7.4.3
eventlet==0.35.0
gunicorn==21.2.0
psutil==5.9.8
redis==5.0.1

# Production monitoring & optimization
accelerate>=0.26.0
bitsandbytes>=0.41.0
